+++ src/app/types/linkType.ts
@@ -1,0 +1,4 @@
+export type LinkType = {
+    name: string,
+    href: string
+}
